% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huff-functions.R
\name{compute_select_prob}
\alias{compute_select_prob}
\title{Title}
\usage{
compute_select_prob(od_matrix, size, exponent_coef)
}
\arguments{
\item{od_matrix}{A dataframe of origin-destination (OD) matrix in long format with column names of "origin", "destination" and "distance"}

\item{size}{A dataframe storing the size (or other measures for "attractiveness")}

\item{exponent_coef}{Impedance coefficient (beta)}
}
\value{
the original `od_matrix` with selection probability of each row computed
}
\description{
Title
}
